requirements.txt
flask==2.2.5
python-telegram-bot==20.7
psycopg2-binary==2.9.3
python-dotenv==1.0.0
apscheduler==3.10.1
nest-asyncio==1.5.6
hypercorn==0.15.0
